# swagger.json
{
  "openapi": "3.0.0",
  "info": { "title": "URL Shortener API", "version": "1.0.0" },
  "servers": [
    { "url": "http://localhost:4000", "description": "Development server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    }
  },
  "security": [],
  "paths": {
    "/signup": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "mysecretpassword"
                  },
                  "name": { "type": "string", "example": "John Doe" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "email": { "type": "string" },
                    "name": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input or user creation failed." }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Log in a user and get a JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "mysecretpassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully, returns JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT authentication token."
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials." }
        }
      }
    },
    "/api/api-keys": {
      "post": {
        "summary": "Create a new API key for the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Optional name for the API key.",
                    "example": "My Marketing Key"
                  },
                  "dailyLimit": {
                    "type": "integer",
                    "description": "Optional daily usage limit for the API key.",
                    "example": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": {
                      "type": "string",
                      "description": "The generated API key."
                    },
                    "id": { "type": "integer" },
                    "createdAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, JWT token missing or invalid."
          }
        }
      }
    },
...
  },
...
}